+++ apps/backend/src/lib/types.tsx
@@ -1,7 +1,3 @@
 import { PrismaClient } from "@prisma/client";
 
 export type PrismaTransaction = Parameters<Parameters<PrismaClient['$transaction']>[0]>[0];
-
-export type Prettify<T> = {
-  [K in keyof T]: T[K];
-} & {};