+++ packages/template/src/index.ts
@@ -1,5 +1,7 @@
 export * from './lib/stack-app';
 
+export { getConvexProvidersConfig } from "./integrations/convex";
+
 // IF_PLATFORM react-like
 export { default as StackHandler } from "./components-page/stack-handler";
 export { useStackApp, useUser } from "./lib/hooks";
@@ -8,6 +10,7 @@ export { StackTheme } from './providers/theme-provider';
 
 export { AccountSettings } from "./components-page/account-settings";
 export { AuthPage } from "./components-page/auth-page";
+export { CliAuthConfirmation } from "./components-page/cli-auth-confirm";
 export { EmailVerification } from "./components-page/email-verification";
 export { ForgotPassword } from "./components-page/forgot-password";
 export { PasswordReset } from "./components-page/password-reset";
@@ -23,5 +26,4 @@ export { OAuthButtonGroup } from "./components/oauth-button-group";
 export { SelectedTeamSwitcher } from "./components/selected-team-switcher";
 export { TeamSwitcher } from "./components/team-switcher";
 export { UserButton } from "./components/user-button";
-export { CliAuthConfirmation } from "./components-page/cli-auth-confirm";
 // END_PLATFORM