+++ vitest.workspace.ts
@@ -3,6 +3,6 @@ import { defineWorkspace } from 'vitest/config';
 export default defineWorkspace([
   'packages/*',
   'apps/*',
-  'examples/*',
   'docs',
+  'examples/*',
 ]);