+++ examples/js-example/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @stackframe/js-example
 
+## 2.8.40
+
+### Patch Changes
+
+- Updated dependencies
+  - @stackframe/js@2.8.40
+
 ## 2.8.39
 
 ### Patch Changes