+++ packages/init-stack/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @stackframe/init-stack
 
+## 2.8.40
+
+### Patch Changes
+
+- Various changes
+- Updated dependencies
+  - @stackframe/stack-shared@2.8.40
+
 ## 2.8.39
 
 ### Patch Changes